requirements.txt
flask
pyjwt
flask-cors
flask-sqlalchemy